requirements.txt
langchain==0.3.17
langchain_community==0.3.16
langchain_core==0.3.33
langchain_openai==0.3.3
faiss-cpu==1.10.0
pandas==2.2.3
pydantic==2.10.6
python-dotenv==1.0.1
twilio~=6.38.1
python-multipart
openpyxl
fastapi 
uvicorn[standard]
pyyaml
rich
selenium
opencv-python